fix phpstan issues

+++ src/Defaults/Trigger/Post/PostPublished.php
@@ -25,7 +25,7 @@ class PostPublished extends PostTrigger {
 	/**
 	 * Post publishing user object
 	 *
-	 * @var \WP_User
+	 * @var \WP_User|false
 	 */
 	public $publishing_user;
 
@@ -87,8 +87,8 @@ class PostPublished extends PostTrigger {
 
 		$this->{ $this->post_type } = $post;
 
-		$this->author          = get_userdata( $this->{ $this->post_type }->post_author );
-		$this->last_editor     = get_userdata( get_post_meta( $this->{ $this->post_type }->ID, '_edit_last', true ) );
+		$this->author          = get_userdata( (int) $this->{ $this->post_type }->post_author );
+		$this->last_editor     = get_userdata( (int) get_post_meta( $this->{ $this->post_type }->ID, '_edit_last', true ) );
 		$this->publishing_user = get_userdata( get_current_user_id() );
 
 		$this->{ $this->post_type . '_creation_datetime' }     = strtotime( $this->{ $this->post_type }->post_date_gmt );
