+++ src/Defaults/Trigger/Post/PostScheduled.php
@@ -18,7 +18,7 @@ class PostScheduled extends PostTrigger {
 	/**
 	 * Post scheduling user object
 	 *
-	 * @var \WP_User
+	 * @var \WP_User|false
 	 */
 	public $scheduling_user;
 
@@ -82,8 +82,8 @@ class PostScheduled extends PostTrigger {
 
 		$scheduling_user_id = get_current_user_id();
 
-		$this->author          = get_userdata( $this->{ $this->post_type }->post_author );
-		$this->last_editor     = get_userdata( get_post_meta( $this->{ $this->post_type }->ID, '_edit_last', true ) );
+		$this->author          = get_userdata( (int) $this->{ $this->post_type }->post_author );
+		$this->last_editor     = get_userdata( (int) get_post_meta( $this->{ $this->post_type }->ID, '_edit_last', true ) );
 		$this->scheduling_user = get_userdata( $scheduling_user_id );
 
 		$this->{ $this->post_type . '_creation_datetime' }     = strtotime( $this->{ $this->post_type }->post_date_gmt );
